+++ R/data.R
@@ -1,4 +1,4 @@
-#' Some animals, their types and their masses
+#' Some animals and their attributes
 #'
 #' @format A data frame with 6 rows and 4 variables:
 #' \describe{
@@ -8,6 +8,17 @@
 #'   \item{mass}{the average mass of the animal in grams} 
 #' }
 "animals"
+
+#' Some animals and their attributes, in rich text
+#'
+#' @format A data frame with 6 rows and 4 variables:
+#' \describe{
+#'   \item{animal}{the name of the animal}, in rich text 
+#'   \item{type}{the type of animal} 
+#'   \item{flies}{whether the animal can fly} 
+#'   \item{mass}{the average mass of the animal in grams} 
+#' }
+"animals_rich"
 
 #' The compositions of some espresso-based beverages
 #'
@@ -33,10 +44,10 @@
 #' @format A data frame with 8 rows and 5 variables:
 #' \describe{
 #'   \item{line}{a line from the poem} 
-#'   \item{xmin}{the xmin coordinate with which to draw the word} 
-#'   \item{xmax}{the xmax coordinate with which to draw the word} 
-#'   \item{ymin}{the ymin coordinate with which to draw the word} 
-#'   \item{ymax}{the ymax coordinate with which to draw the word} 
+#'   \item{xmin}{the xmin coordinate with which to draw the line} 
+#'   \item{xmax}{the xmax coordinate with which to draw the line} 
+#'   \item{ymin}{the ymin coordinate with which to draw the line} 
+#'   \item{ymax}{the ymax coordinate with which to draw the line} 
 #'   \item{linenumber}{the line number} 
 #' }
 "gold"