+++ R/geom_bar_text.R
@@ -68,6 +68,7 @@ GeomBarText <- ggplot2::ggproto(
   ggplot2::Geom,
   required_aes = c("x", "y"),
   default_aes = ggplot2::aes(
+    label = NULL,
     alpha = 1,
     angle = 0,
     colour = "black",